Change build-dist output and support S3 and Github Release deploys
Changed the output of the `dist` folder so all output files are in the
same folder, and now have the type (umd, amd, commonjs, system) in
their filename. The es6 target still has its own folder.

Updated the webpack output to only create a minified version of files
that are used in the browser (normal and umd), because other files will
only be used in a build system that can do the minification.

Added a `package-dist.js` script that prepares the dist for deploys to
s3 and GitHub release. It moves the files in the dist folder to a name +
version folder to mimic the s3 folder structure. It also packages the
es6 folder into a zip file so it can be linked.
For the GitHub release it creates a .zip and .tar.gz file with the
contents of the original dist folder.

Updated the `.travis.yml` to run the `build-dist` and `package-dist`
scripts before the deploy, and added configurations and tokens for the
s3 and Github Release deploys.

Updated the `npm run clean`, `.gitignore` and `.npmignore` to also
clean/ignore the two dist archives in the root.

re #11