Add sourcemap support for coverage by using remap-istanbul
Added remap-istanbul and all supporting karma plugins to correctly
remap the coverage results back to their TypeScript source.

Since all generations happens in memory by Webpack, we needed inline
sourcemap support, provided by the Karma plugins, the Webpack devtool
setting, and the webpack typescript loader.

The normal `ts-loader` doesn't support inline sourcemaps, so we switched
to `awesome-typescript-loader`. We created a new `tsconfig.test.json`
that is the same as the normal `tsconfig.json` but enables inline
sourcemaps by adding `"inlineSourceMap": true`.

The webpack dist builds have also switched to `awesome-typescript-loader`
to keep the configuration consistent.

To make the `awesome-typescript-loader` correctly work with const enums,
we needed to enable `preserveConstEnums` property in the `tsconfig.json`.

We also switched he `test/index.js` to `index.ts` and included
`webpack-env` in the `typings.json` so all wepack requires work in
TypeScript.

To have correct coverage reports consisting of _all_ the source files,
we've added a `Dummy.ts` file in the `src/` folder that is not
referenced by any tests, but must show up in the coverage results.
If this is not the case, we know our test setup is incorrect.

re #15