Fix typescript compiler input files
Create a tsconfig.webpack.json file which we will use for compiling
typescript from webpack. This config will be the same as the
default tsconfig.json, but we will not include the
src/index.ts entry file. Our typescript loader in webpack will
pass the input file to the compiler for us.

In the tsconfig.test.json, use the same files array as
tsconfig.webpack.json. This will also solve some of the globbing
issues we have with awesome-typescript-loader. Fixes #26