Remove the noEmitHelpers from tsconfig.json
The noEmitHelpers will make our bundle size smaller when used in
combination with the ts-helpers module from npm. However, we cannot
use this mechanism in builds where we output a JS file per JS
module. Until we have a way to include ts-helpers conditionally
in single-file output builds, we have to remove this option.