Create separate entry points for different builds
Move all the 'bundle' builds to webpack, and add a `bundle.ts` next to
`index.ts` to export files that are needed for the bundle.
The `bundle.ts` should export all classes, functions and enums that are
exported by other classes in that file, so they are available on the
top-level export, because that's the only way to access them.

Add the files that should be compiled and moved, but are not included
via the `index.ts`, for npm and es6 targets to the `tsconfig.json`
`"files"` section. This can be replaced with the `"include"` section,
that supports globbing, when we switch to TS2, so there is no manual
work here.

Renamed the `webpack.config.js` to `webpack.config.dist.js` to make the
usage more clear. Also changed the entry from an array to a string to
remove the useless 0 webpack module as the 'main' chunk'.
The entry is now pointed to the `bundle.ts` instead of the `index.ts`.

Removed the `typescript-system` task as it didn't really create a
bundle, it just concatenated the files with internal names in the
register call that might conflict when the module is loaded.
Used the `webpack-system-register` plugin that wraps a single
`system.register` around the normal webpack bundle, so it's just one
registered module.

fixes #29