Improve npm scripts and add script comments
Added a cleaner structure to the naming of npm scripts, where the new
format also support the `npm-run-all` sub-script wildcards (e.g.
`npm-run-all clean:*`).

Added a `scriptsComment` block that has a description for each task.
This is the only method that is not removed when a third party script
changes the `package.json` (e.g. `npm i -D foo`).

Update the `.travis.yml` file to reflect the new changes, and explicitly
specify what scripts to run in travis. Now the `test` script runs actual
tests, and the `validate` script does validation + tests, which is
called in travis and with the `pre-push` hook.

Introduced the `shx` (`shelljs` cli wrapper) module so we can use unix
commands in npm scripts that are execute cross-platform in node.
This was needed for moving files (when generating the npm sources),
and also replaces `rimraf`.