Fix sourcemap support for karma
Remove karma-source-map-support dependency, it is no longer needed.
The karma-sourcemap-loader module will handle source maps for us.

Add the sourcemap preprocessor to the test files as well so the
test file line numbers also match.

Instead of the devtool option in the webpack test config, use
the SourceMapDevToolPlugin. This will also generate source
maps on files with a .ts extension.

Fixes #15